Added a bit of docmumentation and gave the whole repo a better strucutre
