+++ README.md
@@ -1,8 +1,7 @@
 # RESEARCH AND INNOVATION LAB
 **Rail Management**
 ---
-An applicaiton dedicated to run the Research and Innovation Lab(RAIL) at Govt.SKSJTI college of Engineering 
+An application dedicated to run the Research and Innovation Lab(RAIL) at Govt.SKSJTI college of Engineering,Bangalore
 ---
 **Developers**
-Tarun Gopalkrishna A
-Prajwal HD
+* Tarun Gopalkrishna A
