+++ json_skeleton_format/communicate.cnf
