--- pythonFiles/testdata.sh
@@ -1,10 +1,0 @@
-echo "REGISTERING STUDENT 'RSK17CS036'"
-./database.py '{"HEADER":{"DATABASE":"rail_test_db","TABLE_NAME":"studs","REQUEST_TYPE":"insert"},"DATA":{"FIELDS":{"usn":"1SK17CS036","rail_id":"RSK17CS036","student_name":"TARUN GOPALKRISHNA A","gender":"M","date_of_birth":"1999-05-01","phone_number":"8296177426","email":"TARUNGOPALKRISHNA@GMAIL.COM","branch":"CS","current_role":"MEMBER"},"SET":null,"WHERE":null},"FOOTER":{"DATA ABOUT THE REQUEST":"reg_stud","COMMENT":"","UPDATE":null,"DEP":null}}'
-echo "REGISTERING STUDENT 'RSK17CS023'"
-./database.py '{"HEADER":{"DATABASE":"rail_test_db","TABLE_NAME":"studs","REQUEST_TYPE":"insert"},"DATA":{"FIELDS":{"usn":"1SK17CS023","rail_id":"RSK17CS023","student_name":"PRAJWAL HD","gender":"M","date_of_birth":"1999-03-03","phone_number":"8951041906","email":"HDPRAJWALGOWDA@GMAIL.COM","branch":"CS","current_role":"TEAM_LEAD"},"SET":null,"WHERE":null},"FOOTER":{"DATA ABOUT THE REQUEST":"reg_stud","COMMENT":"","UPDATE":null,"DEP":null}}'
-echo "REGISTERING TEAM 'RTO19A'"
-./database.py '{"HEADER":{"DATABASE":"rail_test_db","TABLE_NAME":"teams","REQUEST_TYPE":"insert"},"DATA":{"FIELDS":{"team_hash":"RTO19A","team_name":"TEST TEAM xD","number_of_members":"2","date_of_team_creation":"2019-08-08"},"SET":null,"WHERE":null},"FOOTER":{"DATA ABOUT THE REQUEST":"reg_team","COMMENT":"","UPDATE":[{"HEADER":{"DATABASE":"rail_test_db","TABLE_NAME":"studs_and_teams_mapping_table","REQUEST_TYPE":"insert"},"DATA":{"FIELDS":{"team_hash":"RTO19A","rail_id":"RSK17CS023"},"SET":null,"WHERE":null},"FOOTER":{"DATA ABOUT THE REQUEST":"update_map_table","COMMENT":"","UPDATE":null,"DEP":null}},{"HEADER":{"DATABASE":"rail_test_db","TABLE_NAME":"studs_and_teams_mapping_table","REQUEST_TYPE":"insert"},"DATA":{"FIELDS":{"team_hash":"RTO19A","rail_id":"RSK17CS036"},"SET":null,"WHERE":null},"FOOTER":{"DATA ABOUT THE REQUEST":"update_map_table","COMMENT":"","UPDATE":null,"DEP":null}}],"DEP":null}}'
-echo "REGISTERING PROJECT 'RPSO0819AA'"
-./database.py '{"HEADER":{"DATABASE":"rail_test_db","TABLE_NAME":"projs","REQUEST_TYPE":"insert"},"DATA":{"FIELDS":{"team_hash":"RTO19A","project_name":"TEST PROJ xD","project_description":"PROJECT TO TEST THE DATABASE BACK END","mentor":"MD SIR","idea_by":"RSK17CS036","type_of_project":"SOFTWARE","expected_duration":"2 MONTHS","date_start":"2019-08-30","status":"YET TO START","priority":"VERY HIGH","technology_stack":"TESTING SOFTWARES","project_id":"RPSO0819AA","project_lead":"RSK17CS036"},"SET":null,"WHERE":null},"FOOTER":{"DATA ABOUT THE REQUEST":"reg_proj","COMMENT":"","UPDATE":null,"DEP":null}}'
-echo "REGISTERING LOGIN 'RSK17CS036'"
-./database.py '{"HEADER":{"DATABASE":"rail_test_db","TABLE_NAME":"attendance","REQUEST_TYPE":"insert"},"DATA":{"FIELDS":{"rail_id":"RSK17CS036","team_hash":"RTO19A","purpose":"RPSO0819AA"},"SET":null,"WHERE":null},"FOOTER":{"DATA ABOUT THE REQUEST":"stud_login","COMMENT":"","UPDATE":null,"DEP":null}}'
