+++ setPermissions.sh
